Create simple math functions and calc GUI
Created the simple functions that will be used to evaluate the numbers
inputted into the calculator. These include the basic addition, subtraction,
multiplication, and division as well as an 'operate' function which takes
in an operator and then calls the appropriate function. I have also
created a simple layout with buttons to represent the numbers, operators,
and clear and delete buttons.