Create getResult function which calculates input
I have added a getResult function which will take in the input entered
by pressing the buttons, then compute each pair of numbers individually
(for example, if the input is '12 + 8 - 10' it will first add 12 + 8 and
then subtract 10 from that result). The function is dynamic and should
work with any number of operations as long as there is at least two
numbers and an operator. If this is not the case then the function should
return 'ERROR'.