Added helper functions for updating display
I added the helper functions updateDisplay and clearDisplay to clean up
my code. Also, pressing '=' now evaluates the input and returns the
result. The getResult function was also updated such that if the display
is 'ERROR', pressing a button will clear the display and then input the
button pressed, letting users avoid having to manually clear the ERROR
message. The getResult function was also updated to round the result, if
necessary, to 5 decimal places so that the result does not flow off the
screen.