Fix bugs relating to display and calculations
I modified the getResult function to only round when necessary, as it
was returning 9.00000 instead of 9 for example. I also modified this
function to alert the user if they try to divide by 0 and then clear
the input so it doesn't crash the calculator. The updateDisplay function
was also updated to limit the input length to 15 characters. This is to
prevent the user from entering too many numbers otherwise it will overflow
off the calculator screen.