Add delete button functionality
Created a new function, deleteInput, which runs once the 'DELETE' button
is clicked. This button removes the last inputted number or operator (this
also removes the spaces around the operator as well). Useful if the user
makes a mistake and does not wish to clear the whole input to start
again.